+++ resources/leiningen/new/aws_lambda_serverless/test/handler_test.clj
@@ -1,7 +1,7 @@
-(ns {{name}}.handler-test
+(ns {{namespace}}.handler-test
   (:require
     [clojure.test :refer :all]
-    [{{name}}.handler :refer :all]))
+    [{{namespace}}.handler :refer :all]))
 
 (deftest a-test
   (testing "FIXME, I fail."